Refactor: 로그인 api 관련 구조 변경 로그인 관련 타입과 함수 ts 경로를 유지 보수가 용이한 구조로 변경 Resolves: #8
